Rename MessageBuilder to MessageDataBuilder
