+++ src/MessageDataBuilder/DetailsSectionBuilder.php
@@ -2,19 +2,19 @@
 
 declare(strict_types=1);
 
-namespace EdifactParser\MessageBuilder;
+namespace EdifactParser\MessageDataBuilder;
 
 use EdifactParser\Segments\SegmentInterface;
 
-class LineItemsMessageBuilder implements MessageBuilderInterface
+class DetailsSectionBuilder implements BuilderInterface
 {
     private array $builders;
-    private SimpleMessageBuilder $currentBuilder;
+    private SimpleBuilder $currentBuilder;
 
     public function addSegment(SegmentInterface $segment): self
     {
         if ($segment->tag() == 'LIN') {
-            $this->currentBuilder = new SimpleMessageBuilder();
+            $this->currentBuilder = new SimpleBuilder();
             $this->builders[$segment->subId()] = $this->currentBuilder;
         }
